# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Noah Nvidia API
  description: Tenant-scoped supervised virtual employee contracts.
  version: 0.2.0
paths:
...
  /api/v1/bootstrap:
    get:
      summary: Bootstrap
      operationId: bootstrap_api_v1_bootstrap_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Bootstrap Api V1 Bootstrap Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/business:
    get:
      summary: Get Business
      operationId: get_business_api_v1_business_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Business Api V1 Business Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Business
      operationId: update_business_api_v1_business_patch
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Business Api V1 Business Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/conversations:
    get:
      summary: List Conversations
      operationId: list_conversations_api_v1_conversations_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Conversations Api V1 Conversations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Conversation
      operationId: create_conversation_api_v1_conversations_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Conversation Api V1 Conversations Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/actions:
    get:
      summary: List Actions
      operationId: list_actions_api_v1_actions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Actions Api V1 Actions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/services:
    get:
      summary: List Services
      operationId: list_services_api_v1_services_get
      parameters:
      - name: active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Services Api V1 Services Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Service
      operationId: create_service_api_v1_services_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Service Api V1 Services Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/quotes:
    get:
      summary: List Quotes
      operationId: list_quotes_api_v1_quotes_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Quotes Api V1 Quotes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Quote
      operationId: create_quote_api_v1_quotes_post
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Quote Api V1 Quotes Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/ledger/export.csv:
    get:
      summary: Export Ledger
      operationId: export_ledger_api_v1_ledger_export_csv_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/receivables:
    get:
      summary: List Receivables
      operationId: list_receivables_api_v1_receivables_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Receivables Api V1 Receivables Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Receivable
      operationId: create_receivable_api_v1_receivables_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceivableCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Receivable Api V1 Receivables Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/documents:
    post:
      summary: Create Document
      operationId: create_document_api_v1_documents_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Document Api V1 Documents Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: List Documents
      operationId: list_documents_api_v1_documents_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Documents Api V1 Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/calendar/find-slots:
    get:
      summary: Find Calendar Slots
      operationId: find_calendar_slots_api_v1_calendar_find_slots_get
      parameters:
      - name: date
        in: query
        required: false
        schema:
          type: string
          default: '2026-09-05'
          title: Date
      - name: duration_minutes
        in: query
        required: false
        schema:
          type: integer
          maximum: 1440
          minimum: 15
          default: 60
          title: Duration Minutes
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Find Calendar Slots Api V1 Calendar Find Slots Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/google/start:
    post:
      summary: Google Start
      operationId: google_start_api_v1_connections_google_start_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Google Start Api V1 Connections Google Start Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/connections/google/callback:
    get:
      summary: Google Callback
      operationId: google_callback_api_v1_connections_google_callback_get
      parameters:
      - name: state
        in: query
        required: true
        schema:
          type: string
          title: State
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Google Callback Api V1 Connections Google Callback
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    BusinessPatch:
      properties:
        name:
          anyOf:
          - type: string
            maxLength: 160
            minLength: 1
          - type: 'null'
          title: Name
        locale:
          anyOf:
          - type: string
            maxLength: 20
            minLength: 2
          - type: 'null'
          title: Locale
        timezone:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 3
          - type: 'null'
          title: Timezone
        currency:
          anyOf:
          - type: string
            maxLength: 3
            minLength: 3
          - type: 'null'
          title: Currency
        working_hours:
          anyOf:
          - {}
          - type: 'null'
          title: Working Hours
        authority_policy:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Authority Policy
      type: object
      title: BusinessPatch
...
    ConversationCreate:
      properties:
        title:
          type: string
          maxLength: 120
          minLength: 1
          title: Title
          default: Operations desk
      type: object
      title: ConversationCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    QuoteCreate:
      properties:
        contact_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Id
        lines:
          items:
            $ref: '#/components/schemas/QuoteLineCreate'
          type: array
          maxItems: 50
          minItems: 1
          title: Lines
        discount_minor:
          type: integer
          minimum: 0.0
          title: Discount Minor
          default: 0
        valid_days:
          type: integer
          maximum: 90.0
          minimum: 1.0
          title: Valid Days
          default: 7
      type: object
      required:
      - lines
      title: QuoteCreate
...
    ReceivableCreate:
      properties:
        contact_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact Id
        quote_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Quote Id
        amount_due_minor:
          type: integer
          maximum: 1000000000.0
          minimum: 0.0
          title: Amount Due Minor
        due_on:
          anyOf:
          - type: string
          - type: 'null'
          title: Due On
      type: object
      required:
      - amount_due_minor
      title: ReceivableCreate
    ServiceCreate:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
        description:
          type: string
          maxLength: 2000
          title: Description
          default: ''
        price_minor:
          type: integer
          maximum: 1000000000.0
          minimum: 0.0
          title: Price Minor
        duration_minutes:
          type: integer
          maximum: 1440.0
          exclusiveMinimum: 0.0
          title: Duration Minutes
        active:
          type: boolean
          title: Active
          default: true
      type: object
      required:
      - name
      - price_minor
      - duration_minutes
      title: ServiceCreate
...